+++ lib/utils/launch_location.dart
@@ -2,8 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:gap/gap.dart';
-import 'package:stoyco_shared/utils/colors.dart';
+
 import 'package:maps_launcher/maps_launcher.dart';
+import 'package:stoyco_shared/design/colors.dart';
 
 /// Enum representing the type of location launch - either via a search query or using specific coordinates.
 enum LaunchLocationType { query, coordinates }